points calculation implemented into gameview and player detail view game list, improved player detail game list ui

+++ iOS/HeadCoach/LiveGameTableViewCell.swift
@@ -55,7 +55,7 @@ class LiveGameTableViewCell: UITableViewCell{
         }
     }
     
-    func setPlayer(fdplayer: FDPlayer, hcplayer: HCPlayer, pts: UInt32, winner: Bool){
+    func setPlayer(fdplayer: FDPlayer, hcplayer: HCPlayer, pts: Float, winner: Bool){
         position.text = HCPositionUtil.positionToString(hcplayer.position)
         name.text = fdplayer.name
         points.text = String(pts)
